requirements.txt
datasets
numpy
opencv-python-headless
pandas
Pillow
pandas
scikit-learn